Rendu final TP Symfony : Blanchi Melvyn

+++ src/Controller/ArticleController.php
@@ -9,6 +9,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted; // Import nécessaire
 
 final class ArticleController extends AbstractController
 {
@@ -16,6 +17,8 @@ final class ArticleController extends AbstractController
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $article = new Article();
+        $article->setUser($this->getUser()); 
+
         $form = $this->createForm(ArticleType::class, $article);
         $form->handleRequest($request);
 
@@ -43,6 +46,10 @@ final class ArticleController extends AbstractController
     #[Route('admin/article/{id}/edit', name: 'article_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Article $article, EntityManagerInterface $entityManager): Response
     {
+        if ($article->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException("Vous n'êtes pas l'auteur de cet article !");
+        }
+
         $form = $this->createForm(ArticleType::class, $article);
         $form->handleRequest($request);
 
@@ -61,6 +68,10 @@ final class ArticleController extends AbstractController
     #[Route('admin/article/{id}', name: 'article_delete', methods: ['POST'])]
     public function delete(Request $request, Article $article, EntityManagerInterface $entityManager): Response
     {
+        if ($article->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException("Action interdite.");
+        }
+
         if ($this->isCsrfTokenValid('delete' . $article->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($article);
             $entityManager->flush();
